requirements.txt
aiohttp>=3.9.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pytrends>=4.9.0
openai>=1.0.0
python-dotenv>=1.0.0
Django>=5.0.0
django-extensions>=3.2.0
asgiref>=3.7.0
